idea 1: api token consistency scope using flags (#6354)
## About the changes
This PR tries to limit the scope the impact of this change to just
requests from edge. This has to be configured as a constraint in the
feature flag that controls the rollout of this feature.

This is an option to https://github.com/Unleash/unleash/pull/6356 which
is more flexible and can be used for other purposes.