fix: admin token requests does not automatically have id (#6501) (#6513)
To check that users do indeed have permissions to update the roles from
project-service, we've been depending on req.user.id. We had one error
on Friday March 8th, where we managed to send undefined/null to a method
that requires a number. This PR assumes that if we have an API token,
and we have admin permissions and userId is not set we're a legacy admin
token.

It uses the util method for extractUserId(req: IAuthRequest |
IApiRequest), so if we've passed through the apiTokenMiddleware first,
we'll have userId -42, if we haven't, we'll get -1337.