chore: automate release actions (#7070)
## About the changes
Automate some of our release steps in the same workflow. This gives us
the ability to opt-out some of the jobs which allows us to release
patches to older versions without setting that patch as the latest
version (currently not covering docker publish but if this test works
successfully we might include it as well).

This comes from tests in our enterprise repository.